back/requirements.txt
fastapi==0.115.5
mangum==0.17.0
uvicorn==0.30.1
PyJWT==2.9.0
